Move util functions to utils directory
