+++ utils/formatMoney.ts
@@ -1,0 +1,9 @@
+/** Formats the film's budget/revenue as a descriptive string. */
+export const formatMoney = (money: number): string => {
+  const formatter: Intl.NumberFormat = new Intl.NumberFormat('en-US', {
+    style: 'currency',
+    currency: 'USD',
+    minimumFractionDigits: 0,
+  });
+  return formatter.format(money);
+};
