+++ utils/formatTime.ts
@@ -1,0 +1,6 @@
+/** Formats the movie runtime as a descriptive string. */
+export const formatTime = (time: number): string => {
+  const hours: number = Math.floor(time / 60);
+  const mins: number = time % 60;
+  return `${hours}h ${mins}m`;
+};
